# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Steak Kenangan Restaurant POS API
  description: |
    REST API for Steak Kenangan Restaurant Point-of-Sale system.

    ## Features
    - Role-based access control (RBAC)
    - JWT authentication
    - Rate limiting
    - Comprehensive order management
    - Inventory tracking
    - Real-time kitchen display
    - Multi-payment processing

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## Roles
    - **admin**: Full system access
    - **manager**: Business operations and reporting
    - **server**: Dine-in orders only
    - **counter**: All order types + payments
    - **kitchen**: Order preparation workflow

    ## Rate Limiting
    - Public endpoints: 30 requests/minute per IP
    - Auth endpoints: 5 requests/minute (brute-force protection)
    - Contact form: 3 requests/5 minutes

    ## Response Format
    All endpoints return standard response format:
    ```json
    {
      "success": boolean,
      "message": string,
      "data": object | array,
      "error": string (optional)
    }
    ```
  version: 1.0.0
  contact:
    name: Steak Kenangan API Support
    email: tech@steakkenangan.com
    url: https://steakkenangan.com
  license:
    name: Proprietary
    url: https://steakkenangan.com/legal/terms.html
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
    APIResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if request was successful
        message:
          type: string
          description: Human-readable message
        data:
          description: Response payload (varies by endpoint)
        error:
          type: string
          description: Error message (only present on failure)
...
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_number:
          type: string
        user_id:
          type: string
          format: uuid
        table_id:
          type: string
          format: uuid
        order_type:
          type: string
          enum: [dine_in, takeaway, delivery]
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, served, completed, cancelled]
        customer_name:
          type: string
        customer_phone:
          type: string
        delivery_address:
          type: string
        subtotal:
          type: number
          format: decimal
        tax_amount:
          type: number
          format: decimal
        discount_amount:
          type: number
          format: decimal
        total_amount:
          type: number
          format: decimal
        notes:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Table:
      type: object
      properties:
        id:
          type: string
          format: uuid
        table_number:
          type: string
        seating_capacity:
          type: integer
        location:
          type: string
        is_occupied:
          type: boolean
        qr_code:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string
paths:
...
  /public/contact:
    post:
      tags:
        - Public
      summary: Submit contact form
      description: |
        Submit contact/inquiry form from public website.

        **Rate Limit**: 3 requests/5 minutes per IP (spam protection)
      operationId: submitContactForm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - subject
                - message
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                phone:
                  type: string
                subject:
                  type: string
                  enum: [general, reservation, catering, feedback, complaint]
                message:
                  type: string
      responses:
        '200':
          description: Contact form submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /customer/table/{qr_code}:
    get:
      tags:
        - Customer
      summary: Get table by QR code
      description: Returns table information for customer self-ordering
      operationId: getTableByQRCode
      parameters:
        - name: qr_code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Table info retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Table'
        '404':
          description: Table not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /customer/orders:
    post:
      tags:
        - Customer
      summary: Create customer order
      description: Create order from customer self-ordering (QR code scan)
      operationId: createCustomerOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - table_id
                - items
              properties:
                table_id:
                  type: string
                  format: uuid
                customer_name:
                  type: string
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      product_id:
                        type: string
                        format: uuid
                      quantity:
                        type: integer
                      special_instructions:
                        type: string
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Order'

  # ==========================================
  # PROFILE ENDPOINTS
  # ==========================================
...